fix(tasks): split assignee resolution and update web test expectations
