+++ apps/web/src/app/(app)/tasks/components/task-form.test.tsx
@@ -1098,6 +1098,7 @@ describe("TaskForm", () => {
     expect(createTaskMock).toHaveBeenCalledWith({
       description: "Build landing page",
       assigneeId: "coworker-2",
+      assigneeOrchestratorId: null,
       context: {
         brand: { enabled: true, source: "default", custom: null },
         briefingEnabled: true,
@@ -1143,6 +1144,7 @@ describe("TaskForm", () => {
     expect(createTaskMock).toHaveBeenCalledWith({
       description: "Build landing page",
       assigneeId: "coworker-2",
+      assigneeOrchestratorId: null,
       context: {
         brand: { enabled: false, source: "default", custom: null },
         briefingEnabled: true,
@@ -1190,6 +1192,7 @@ describe("TaskForm", () => {
     expect(createTaskMock).toHaveBeenCalledWith({
       description: "Build landing page",
       assigneeId: "coworker-2",
+      assigneeOrchestratorId: null,
       context: {
         brand: {
           enabled: true,
@@ -1809,6 +1812,7 @@ describe("TaskForm", () => {
     expect(onCreateTask).toHaveBeenCalledWith({
       description: "Write docs",
       assigneeId: "coworker-2",
+      assigneeOrchestratorId: null,
       context: {
         brand: { enabled: true, source: "default", custom: null },
         briefingEnabled: true,
